fixed a bug in rank calculation

+++ src/main/java/org/la4j/factory/CCSFactory.java
@@ -160,7 +160,7 @@ public class CCSFactory extends CompressedFactory implements Factory {
         int columnPointers[] = new int[size + 1];
 
         for (int i = 0; i < size; i++) {
-            values[i] = (double) 1.0;
+            values[i] = 1.0;
             rowIndices[i] = i;
             columnPointers[i] = i;
         }
